xcomponent add ndk for key event

+++ frameworks/bridge/declarative_frontend/engine/functions/js_key_function.cpp
@@ -30,7 +30,7 @@ JSRef<JSObject> JsKeyFunction::createKeyEvent(KeyEventInfo& event)
     keyEventObj->SetProperty<int32_t>("keySource", static_cast<int32_t>(event.GetKeySource()));
     keyEventObj->SetProperty<int64_t>("deviceId", event.GetDeviceId());
     keyEventObj->SetProperty<int32_t>("metaKey", event.GetMetaKey());
-    keyEventObj->SetProperty<float>("timestamp", event.GetTimeStamp().time_since_epoch().count());
+    keyEventObj->SetProperty<double>("timestamp", static_cast<double>(event.GetTimeStamp().time_since_epoch().count()));
     keyEventObj->SetPropertyObject("stopPropagation", JSRef<JSFunc>::New<FunctionCallback>(JsStopPropagation));
     keyEventObj->Wrap<KeyEventInfo>(&event);
     return keyEventObj;
